se modularizan las validaciones handlers en el archivo validations.py
